Separate SiestaUI framework, anticipating Linux support
