+++ Source/Siesta/Support/Siesta-ObjC.swift
@@ -299,38 +299,6 @@ public extension Resource
         }
     }
 
-#if !os(OSX)
-extension ResourceStatusOverlay: _objc_ResourceObserver
-    {
-    public func resourceChanged(resource: Resource, event eventString: String)
-        {
-        if let event = ResourceEvent.fromDescription(eventString)
-            { resourceChanged(resource, event: event) }
-        }
-    }
-
-extension ResourceStatusOverlay
-    {
-    @objc(displayPriority)
-    public var _objc_displayPriority: [String]
-        {
-        get {
-            return displayPriority.map { $0.rawValue }
-            }
-
-        set {
-            displayPriority = newValue.flatMap
-                {
-                let condition = ResourceStatusOverlay.StateRule(rawValue: $0)
-                if condition == nil
-                    { Swift.print("WARNING: ignoring unknown ResourceStatusOverlay.StateRule \"\($0)\"") }
-                return condition
-                }
-            }
-        }
-    }
-#endif
-
 public extension Resource
     {
     private func _objc_wrapRequest(
