+++ Source/SiestaUI/ResourceStatusOverlay.swift
@@ -6,6 +6,7 @@
 //  Copyright © 2016 Bust Out Solutions. All rights reserved.
 //
 
+import Siesta
 import Foundation
 import CoreGraphics
 import UIKit
